DRTII-1856 Move more classes to drt-lib
