+++ server/src/main/scala/services/liveviews/PassengersLiveView.scala
@@ -11,9 +11,9 @@ import org.apache.pekko.util.Timeout
 import drt.shared.CrunchApi.{MinutesContainer, PassengersMinute}
 import drt.shared.{CodeShares, CrunchApi}
 import org.slf4j.LoggerFactory
-import slickdb.AggregatedDbTables
 import uk.gov.homeoffice.drt.actor.state.ArrivalsState
 import uk.gov.homeoffice.drt.arrivals.{ApiFlightWithSplits, Arrival}
+import uk.gov.homeoffice.drt.db.AggregatedDbTables
 import uk.gov.homeoffice.drt.db.dao.{CapacityHourlyDao, PassengersHourlyDao}
 import uk.gov.homeoffice.drt.db.serialisers.{CapacityHourlySerialiser, PassengersHourlySerialiser}
 import uk.gov.homeoffice.drt.db.tables.{CapacityHourly, PassengersHourly, PassengersHourlyRow}
