+++ server/src/main/scala/services/liveviews/QueuesLiveView.scala
@@ -2,7 +2,7 @@ package services.liveviews
 
 import drt.shared.CrunchApi.CrunchMinutes
 import org.slf4j.LoggerFactory
-import slickdb.AggregatedDbTables
+import uk.gov.homeoffice.drt.db.AggregatedDbTables
 import uk.gov.homeoffice.drt.db.dao.QueueSlotDao
 import uk.gov.homeoffice.drt.models.CrunchMinute
 import uk.gov.homeoffice.drt.ports.PortCode
