+++ server/src/main/scala/slickdb/DropIns.scala
@@ -4,6 +4,7 @@ import drt.shared.DropIn
 import org.joda.time.DateTime
 import slick.jdbc.PostgresProfile.api._
 import slick.lifted.ProvenShape
+import uk.gov.homeoffice.drt.db.AggregatedDbTables
 
 import java.sql.Timestamp
 import java.time.format.DateTimeFormatter
