+++ server/src/main/scala/slickdb/FeatureGuideViewRow.scala
@@ -2,6 +2,7 @@ package slickdb
 
 import slick.jdbc.PostgresProfile.api._
 import slick.lifted.ProvenShape
+import uk.gov.homeoffice.drt.db.AggregatedDbTables
 
 import java.sql.Timestamp
 import scala.concurrent.{ExecutionContext, Future}
