+++ server/src/main/scala/uk/gov/homeoffice/drt/testsystem/TestTables.scala
@@ -1,16 +1,16 @@
 package uk.gov.homeoffice.drt.testsystem
 
 import actors.DrtParameters
-import org.apache.pekko.stream.scaladsl.Source
 import com.google.inject.Inject
-import drt.shared.{DropIn, Shift, UserPreferences}
+import drt.shared.{DropIn, Shift}
 import manifests.passengers.{BestAvailableManifest, ManifestPaxCount}
 import manifests.{ManifestLookupLike, UniqueArrivalKey}
-import play.api.Configuration
+import org.apache.pekko.stream.scaladsl.Source
 import slickdb._
 import uk.gov.homeoffice.drt.arrivals.VoyageNumber
 import uk.gov.homeoffice.drt.db.dao.{IABFeatureDao, IUserFeedbackDao}
-import uk.gov.homeoffice.drt.db.tables.{ABFeatureRow, UserFeedbackRow}
+import uk.gov.homeoffice.drt.db.tables.{ABFeatureRow, UserFeedbackRow, UserRow, UserTableLike}
+import uk.gov.homeoffice.drt.models.UserPreferences
 import uk.gov.homeoffice.drt.ports.PortCode
 import uk.gov.homeoffice.drt.service.staffing.ShiftsService
 import uk.gov.homeoffice.drt.time.{SDate, SDateLike}
