+++ server/src/test/scala/manifests/ManifestLookupSpec.scala
@@ -3,15 +3,15 @@ package manifests
 import manifests.passengers.BestAvailableManifest
 import org.specs2.specification.Before
 import services.crunch.CrunchTestLike
-import slickdb.{ProcessedJsonRow, ProcessedZipRow, VoyageManifestPassengerInfoRow}
 import uk.gov.homeoffice.drt.Nationality
 import uk.gov.homeoffice.drt.arrivals.EventTypes.DC
 import uk.gov.homeoffice.drt.arrivals.{CarrierCode, VoyageNumber}
+import uk.gov.homeoffice.drt.db.AggregateDbH2
+import uk.gov.homeoffice.drt.db.tables.{ProcessedJsonRow, ProcessedZipRow, VoyageManifestPassengerInfoRow}
 import uk.gov.homeoffice.drt.models.DocumentType.Passport
 import uk.gov.homeoffice.drt.models.ManifestPassengerProfile
 import uk.gov.homeoffice.drt.ports.SplitRatiosNs.SplitSources.Historical
 import uk.gov.homeoffice.drt.ports.{PaxAge, PortCode}
-import uk.gov.homeoffice.drt.testsystem.db.AggregateDbH2
 import uk.gov.homeoffice.drt.time.SDate
 
 import java.sql.Timestamp
