--- shared/src/main/scala/drt/shared/UserPreferences.scala
@@ -1,12 +1,0 @@
-package drt.shared
-
-import upickle.default._
-
-case class UserPreferences(userSelectedPlanningTimePeriod: Int,
-                           hidePaxDataSourceDescription: Boolean,
-                           showStaffingShiftView: Boolean,
-                           desksAndQueuesIntervalMinutes: Int)
-
-object UserPreferences {
-  implicit val rw: ReadWriter[UserPreferences] = macroRW
-}
